Refactorizar mas codigo.
